requirements.txt
google-genai
streamlit
pydantic